split model sub-package into smaller modules, add makefile-vars for these
