+++ Makefile
@@ -35,6 +35,7 @@ include sbin/makefile-vars
 include ermrest/makefile-vars
 include ermrest/exception/makefile-vars
 include ermrest/ermpath/makefile-vars
+include ermrest/model/makefile-vars
 include ermrest/url/makefile-vars
 include ermrest/url/ast/makefile-vars
 include ermrest/url/ast/data/makefile-vars
